fold PyrenewWastewaterData into PyrenewHEWData and update test
