+++ tests/test_pyrenew_hew_data.py
@@ -1,9 +1,10 @@
 import datetime
 
+import numpy as np
+import polars as pl
 import pytest
 
 from pyrenew_hew.pyrenew_hew_data import PyrenewHEWData
-from pyrenew_hew.pyrenew_wastewater_data import PyrenewWastewaterData
 
 
 @pytest.mark.parametrize(
@@ -97,15 +98,12 @@ def test_to_forecast_data(
     """
     Test the to_forecast_data method
     """
-    ww_dat = PyrenewWastewaterData()
-
     data = PyrenewHEWData(
         n_ed_visits_data_days=n_ed_visits_data_days,
         n_hospital_admissions_data_days=n_hospital_admissions_data_days,
         first_ed_visits_date=first_ed_visits_date,
         first_hospital_admissions_date=first_hospital_admissions_date,
         right_truncation_offset=right_truncation_offset,
-        **ww_dat.to_pyrenew_hew_data_args(),
     )
 
     assert data.right_truncation_offset == right_truncation_offset
@@ -133,3 +131,87 @@ def test_to_forecast_data(
 
     assert forecast_data.first_wastewater_date == data.first_data_date_overall
     assert forecast_data.data_observed_disease_wastewater_conc is None
+
+
+def test_pyrenew_wastewater_data():
+    first_training_date = datetime.date(2023, 1, 1)
+    last_training_date = datetime.date(2023, 7, 23)
+    dates = pl.date_range(
+        first_training_date,
+        last_training_date,
+        interval="1w",
+        closed="both",
+        eager=True,
+    )
+
+    ww_raw = pl.DataFrame(
+        {
+            "date": dates.extend(dates),
+            "site": [200] * 30 + [100] * 30,
+            "lab": [21] * 60,
+            "lab_site_index": [1] * 30 + [2] * 30,
+            "site_index": [1] * 30 + [2] * 30,
+            "log_genome_copies_per_ml": np.log(
+                np.abs(np.random.normal(loc=500, scale=50, size=60))
+            ),
+            "log_lod": np.log([20] * 30 + [15] * 30),
+            "site_pop": [200_000] * 30 + [400_000] * 30,
+        }
+    )
+
+    ww_data = ww_raw.with_columns(
+        (pl.col("log_genome_copies_per_ml") <= pl.col("log_lod"))
+        .cast(pl.Int8)
+        .alias("below_lod")
+    )
+
+    first_ed_visits_date = datetime.date(2023, 1, 1)
+    first_hospital_admissions_date = datetime.date(2023, 1, 7)  # Saturday
+    first_wastewater_date = datetime.date(2023, 1, 1)
+    n_forecast_points = 10
+
+    data = PyrenewHEWData(
+        first_ed_visits_date=first_ed_visits_date,
+        first_hospital_admissions_date=first_hospital_admissions_date,
+        first_wastewater_date=first_wastewater_date,
+        nwss_training_data=ww_data,
+        population_size=1e6,
+    )
+
+    forecast_data = data.to_forecast_data(n_forecast_points)
+    assert forecast_data.data_observed_disease_wastewater_conc is None
+    assert data.data_observed_disease_wastewater_conc is not None
+
+    assert forecast_data.ww_censored is None
+    assert data.ww_censored is not None
+
+    assert forecast_data.ww_uncensored is None
+    assert data.ww_uncensored is not None
+
+    assert forecast_data.ww_log_lod is None
+    assert data.ww_log_lod is not None
+
+    assert forecast_data.ww_observed_lab_sites is None
+    assert data.ww_observed_lab_sites is not None
+
+    assert forecast_data.ww_observed_subpops is None
+    assert data.ww_observed_subpops is not None
+
+    assert data.model_t_obs_wastewater is not None
+    assert forecast_data.model_t_obs_wastewater is None
+
+    assert np.array_equal(data.n_ww_lab_sites, forecast_data.n_ww_lab_sites)
+    assert np.array_equal(data.pop_fraction, forecast_data.pop_fraction)
+
+    assert np.array_equal(
+        data.data_observed_disease_wastewater_conc,
+        ww_data["log_genome_copies_per_ml"],
+    )
+    assert len(data.ww_censored) == len(
+        ww_data.filter(pl.col("below_lod") == 1)
+    )
+    assert len(data.ww_uncensored) == len(
+        ww_data.filter(pl.col("below_lod") == 0)
+    )
+    assert np.array_equal(data.ww_log_lod, ww_data["log_lod"])
+    assert data.n_ww_lab_sites == ww_data["lab_site_index"].n_unique()
